Extract shared E2E code into packages/e2e
