+++ apps/web/package.json
@@ -60,6 +60,7 @@
   "devDependencies": {
     "@eslint/js": "^9.39.2",
     "@faker-js/faker": "^10.3.0",
+    "@smela/e2e": "workspace:*",
     "@hookform/devtools": "^4.4.0",
     "@playwright/test": "^1.58.2",
     "@sentry/vite-plugin": "^4.9.1",
@@ -84,7 +85,6 @@
     "eslint-plugin-storybook": "^10.2.10",
     "globals": "^17.0.0",
     "lint-staged": "^16.2.6",
-    "mailisk": "^2.2.2",
     "prettier": "^3.6.2",
     "prop-types": "^15.8.1",
     "rollup-plugin-visualizer": "^6.0.5",
