+++ packages/e2e/src/actions/logout.js
