+++ packages/e2e/src/actions/signup.js
@@ -1,13 +1,13 @@
 export const fillSignupFormAndSubmit = async (
   page,
   { firstName, lastName, email, password },
-  t
+  t,
 ) => {
   const firstNameInput = page.getByLabel(t.firstName.label)
   const lastNameInput = page.getByLabel(t.lastName.label)
   const emailInput = page.getByLabel(t.email.label)
   const passwordInput = page.getByRole('textbox', {
-    name: t.password.label.default
+    name: t.password.label.default,
   })
   const signupButton = page.getByRole('button', { name: t.signUp })
 
@@ -22,6 +22,6 @@ export const fillSignupFormAndSubmit = async (
     lastNameInput,
     emailInput,
     passwordInput,
-    signupButton
+    signupButton,
   }
 }
